Refactor exception package structure
