+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/ContractUtil.java
@@ -2,6 +2,10 @@ package de.upb.cs.uc4.chaincode.contract;
 
 import de.upb.cs.uc4.chaincode.contract.approval.ApprovalContractUtil;
 import de.upb.cs.uc4.chaincode.exceptions.*;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.ledgeraccess.LedgerStateNotFoundError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.ledgeraccess.UnprocessableLedgerStateError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.LedgerAccessError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.ValidationError;
 import de.upb.cs.uc4.chaincode.model.ApprovalList;
 import de.upb.cs.uc4.chaincode.model.errors.DetailedError;
 import de.upb.cs.uc4.chaincode.model.errors.GenericError;
