+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/admission/AdmissionContractUtil.java
@@ -1,7 +1,7 @@
 package de.upb.cs.uc4.chaincode.contract.admission;
 
-import de.upb.cs.uc4.chaincode.exceptions.LedgerAccessError;
-import de.upb.cs.uc4.chaincode.exceptions.ParameterError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.LedgerAccessError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.ParameterError;
 import de.upb.cs.uc4.chaincode.model.*;
 import de.upb.cs.uc4.chaincode.model.errors.InvalidParameter;
 import de.upb.cs.uc4.chaincode.contract.ContractUtil;
