+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/certificate/CertificateContract.java
@@ -1,7 +1,7 @@
 package de.upb.cs.uc4.chaincode.contract.certificate;
 
 import de.upb.cs.uc4.chaincode.contract.ContractBase;
-import de.upb.cs.uc4.chaincode.exceptions.ParameterError;
+import de.upb.cs.uc4.chaincode.exceptions.serializable.ParameterError;
 import org.hyperledger.fabric.contract.Context;
 import org.hyperledger.fabric.contract.annotation.Contract;
 import org.hyperledger.fabric.contract.annotation.Transaction;
