+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/exceptions/serializable/LedgerAccessError.java
@@ -1,4 +1,6 @@
-package de.upb.cs.uc4.chaincode.exceptions;
+package de.upb.cs.uc4.chaincode.exceptions.serializable;
+
+import de.upb.cs.uc4.chaincode.exceptions.SerializableError;
 
 public abstract class LedgerAccessError extends SerializableError {
