+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/exceptions/serializable/ledgeraccess/UnprocessableLedgerStateError.java
@@ -1,4 +1,6 @@
-package de.upb.cs.uc4.chaincode.exceptions;
+package de.upb.cs.uc4.chaincode.exceptions.serializable.ledgeraccess;
+
+import de.upb.cs.uc4.chaincode.exceptions.serializable.LedgerAccessError;
 
 public class UnprocessableLedgerStateError extends LedgerAccessError {
     public UnprocessableLedgerStateError(String jsonError) {
